feat(app): implement robust connection (re)establishment and guard memory usage
Retry failed connection initializations so that applications can start even when no connection is available to RabbitMQ.
Implement tracking for KBs of messages processed for both consumers and producers.
Buffer published messages when the connection to RabbitMQ is down and flush them when the connection is up again.
Offer configuration options for message accounting and buffering.
Implement callbacks for a configured hibernation threshold and for when a message is buffered and for when buffered messages can be flushed.